% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{clean_bt_list_db}
\alias{clean_bt_list_db}
\title{Clean BabyTime List DB File}
\usage{
clean_bt_list_db(list_db)
}
\arguments{
\item{list_db}{output from `read_one_bt_activity_file`}
}
\value{
a Clean BT List DB object
}
\description{
Clean BabyTime List DB File
}
\examples{
data_dir <- system.file('extdata', package = 'babyTimeR')
baby_dann_db <- read_one_bt_activity_file(
  infile = file.path(data_dir, 'activity_Dann_202411.txt')
) |>
    clean_bt_list_db()
names(baby_dann_db)
}
