% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{combine_clean_bt_list_dbs}
\alias{combine_clean_bt_list_dbs}
\title{Combine two 'Babylist' List DBs}
\usage{
combine_clean_bt_list_dbs(la, lb)
}
\arguments{
\item{la}{a Clean BT List DB object}

\item{lb}{another Clean BT List DB object}
}
\value{
a Clean BT List DB object that contains all elements from `la` and `lb`
}
\description{
Generally not needed if you parse all files at once, but can be useful if
combining data extracted at multiple times or across multiple directories
}
\examples{
\dontrun{
baby_dann_db_may <- read_one_bt_activity_file(
    infile = file.path('Data', 'activity_Dann_202305.zip')
) |>
clean_bt_list_db()
baby_dann_db_june <- read_one_bt_activity_file(
    infile = file.path('Data', 'activity_Dann_202306.zip')
) |>
clean_bt_list_db()
baby_dann_db <- combine_clean_bt_list_dbs(
    baby_dann_db_may,
    baby_dann_db_june
)
}
}
